+++ MASFoundation/Classes/_private_/services/access/MASAccessService.h
@@ -24,7 +24,7 @@ typedef NS_ENUM(NSInteger, MASAccessValueType)
 {
     MASAccessValueTypeUknonw = -1,
     MASAccessValueTypeAccessToken,
-    MASAccessValueTypeAuthenticatedUsername,
+    MASAccessValueTypeAuthenticatedUserObjectId,
     MASAccessValueTypeRefreshToken,
     MASAccessValueTypeScope,
     MASAccessValueTypeTokenType,