+++ ChangeLog.md
@@ -4,6 +4,7 @@
  
 - Adding listAttibutes to MASObject in MASFoundation framework after removing from MASStorage MASObject category.
 - Fix the bug where MASUser object's isAuthenticated property returns wrong value for other user objects retrieved from MASIdentityManagement.
+- Fix the bug where isCurrentUser flag was not determined dynamically when retrieved from MASIdentityManagement.  MASFoundation will determine isCurrentUser flag dynamically by [MASUser currentUser] and/or user retrival from MASIdentityManagement.
 - Fix the bug when passing "cancel" as YES into MASBasicCredentialBlock, SDK was failing to notify the original caller of the request; it was only notifying the completion block of MASBasicCredentialBlock.
 - Fix the bug where version number and version string were returning incorrect values.